Move withShoppingCart tests to own file
